requirements.txt
qcloudapi-sdk-python
netaddr
